minor changes. working on project now

+++ src/main/java/com/prjmng/services/TeamService.java
@@ -16,8 +16,8 @@ import com.prjmng.shared.DTOs.teams.CreateTeamRequest;
 import com.prjmng.shared.DTOs.teams.TeamMemberResponse;
 import com.prjmng.shared.DTOs.teams.TeamResponse;
 import com.prjmng.shared.DTOs.users.UserResponse;
+import jakarta.persistence.EntityNotFoundException;
 import jakarta.validation.Valid;
-import jakarta.ws.rs.NotFoundException;
 import lombok.RequiredArgsConstructor;
 import org.jspecify.annotations.NonNull;
 import org.springframework.data.domain.Page;
@@ -37,7 +37,7 @@ public class TeamService {
     private final UserRepository userRepository;
 
     public TeamResponse createTeam(@Valid CreateTeamRequest createTeamRequest, UUID userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new NotFoundException("User with this " + userId + " id was not foud"));
+        User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User with this " + userId + " id was not foud"));
 
         Organization organization = organizationRepository.findByIdAndOwnerId(createTeamRequest.getOrgId(), user.getId())
                 .orElseThrow(() -> new RuntimeException("User is not owner and cannot create team in this organization"));
@@ -68,6 +68,12 @@ public class TeamService {
         teamRepository.deleteById(teamId);
     }
 
+    public Page<TeamResponse> getAllWithPagination(PageRequest pageRequest, UUID userId, UUID orgId, String name) {
+        Page<Team> teams = teamMemberRepository.findTeamsByUserId(userId, name, orgId, pageRequest);
+
+        return teams.map(t -> mapToResponse(t, t.getOrganization()));
+    }
+
     public Page<TeamResponse> getAllWithPagination(PageRequest pageRequest, UUID orgId, String name) {
         Specification<Team> specification = TeamSpecifications.hadOrgId(orgId)
                 .and(TeamSpecifications.hasName(name));
@@ -77,8 +83,14 @@ public class TeamService {
         return teams.map(t -> mapToResponse(t, t.getOrganization()));
     }
 
+    public TeamResponse getById(UUID id, UUID userId) {
+        TeamMember teamMember = teamMemberRepository.findTeamByUserIdAndTeamId(userId, id).orElseThrow(() -> new EntityNotFoundException("Team member with " + id + " id was not found"));
+        Team team = teamMember.getTeam();
+        return mapToResponse(team, team.getOrganization());
+    }
+
     public TeamResponse getById(UUID id) {
-        Team team = teamRepository.findById(id).orElseThrow(() -> new NotFoundException("Team with " + id + " id was not found"));
+        Team team = teamRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Team with " + id + " id was not found"));
 
         return mapToResponse(team, team.getOrganization());
     }
@@ -89,8 +101,33 @@ public class TeamService {
             throw new RuntimeException("User already is member");
         }
 
-        Team team = teamRepository.findById(id).orElseThrow(() -> new NotFoundException("Team with " + id + " id was not found"));
-        User user = userRepository.findById(request.getUserId()).orElseThrow(() -> new NotFoundException("User with " + request.getUserId() + " id was not found"));
+        Team team = teamRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Team with " + id + " id was not found"));
+        User user = userRepository.findById(request.getUserId()).orElseThrow(() -> new EntityNotFoundException("User with " + request.getUserId() + " id was not found"));
+
+        TeamMember member = TeamMember
+                .builder()
+                .userId(request.getUserId())
+                .team(team)
+                .role(request.getRole())
+                .build();
+
+        member = teamMemberRepository.save(member);
+
+        return mapToResponse(member, user);
+    }
+
+    public TeamMemberResponse addMember(UUID id, @Valid CreateTeamMemberRequest request, UUID userId) {
+        boolean isUserJustMember = teamMemberRepository.existsByTeamIdAndUserIdAndRole(id, userId, TeamMemberRole.MEMBER);
+        if(isUserJustMember) {
+            throw new RuntimeException("User cannot add new member to team");
+        }
+        boolean isAlreadyMember = teamMemberRepository.existsByUserIdAndTeamId(request.getUserId(), id);
+        if(isAlreadyMember) {
+            throw new RuntimeException("User already is member");
+        }
+
+        Team team = teamRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Team with " + id + " id was not found"));
+        User user = userRepository.findById(request.getUserId()).orElseThrow(() -> new EntityNotFoundException("User with " + request.getUserId() + " id was not found"));
 
         TeamMember member = TeamMember
                 .builder()
@@ -107,13 +144,16 @@ public class TeamService {
 
     public void removeMember(UUID teamId, UUID memberId, UUID userId) {
         if (!teamRepository.existsById(teamId))
-            throw new NotFoundException("Team with " + teamId + " id was not found");
+            throw new EntityNotFoundException("Team with " + teamId + " id was not found");
 
         boolean isUserJustMember = teamMemberRepository.existsByTeamIdAndUserIdAndRole(teamId, userId, TeamMemberRole.MEMBER);
         if(isUserJustMember) {
             throw new RuntimeException("Member role cannot remove team-member from team");
         }
-        TeamMember member = teamMemberRepository.findById(memberId).orElseThrow(() -> new NotFoundException("Team member with " + memberId + " id was not found"));
+        TeamMember member = teamMemberRepository.findById(memberId).orElseThrow(() -> new EntityNotFoundException("Team member with " + memberId + " id was not found"));
+        if(member.getRole() == TeamMemberRole.OWNER) {
+            throw new RuntimeException("Owner cannot be removed from team");
+        }
         teamMemberRepository.delete(member);
     }
